Update changeRedirectCode method in InertiaMiddleware
Continue the improvement of the library based on Inertia Laravel Library and what the documentation states. A 409 with a X-Inertia-Location Header should produce an external redirect in the inertia.js library. But for this to happen we need to remove the X-Inertia header, if not, inertia.js won't handle the external redirect.

Add a test for this path inside changeRedirectCode method of the InertiaMiddleware