Implement Inertia method location
Laravel Inertia Library has the option to redirect to external url or routes not handled with Inertia, via the location method.

In this method we add the possibility to send the destination as a ResponseInterface, in case we already have a response or a string so a url as a string can be sent.

Status Code option is also added in order to provide the developer the ability to add a 303 if required.